Přidání ukládání času přidáni ukolu do localstorage, oprava editBtn aby vyvolal modal, zmenšený prostor mezi ikony, přidaná ikona pro čas
